concurrency/concurrency.go
package concurrency

type WebsiteChecker func(string) bool
type CheckFunction func(WebsiteChecker, []string) map[string]bool
type result struct {
	url   string
	check bool
}

func CheckWebsite(wc WebsiteChecker, urls []string) map[string]bool {
	results := make(map[string]bool)
	resultChannel := make(chan result)
	for _, url := range urls {
		go func(u string) {
			resultChannel <- result{u, wc(u)}
		}(url)
	}
	for i := 0; i < len(urls); i++ {
		r := <-resultChannel
		results[r.url] = r.check
	}
	return results
}

func CheckWebsite1(wc WebsiteChecker, urls []string) map[string]bool {
	results := make(map[string]bool)
	for _, url := range urls {
		results[url] = wc(url)
	}
	return results
}
